separated manual test to enable ci to work
