+++ django/researchdata/testdata/README.md
@@ -1,0 +1,4 @@
+# README
+
+- This dir contains text files that include sample data outputted by the CWB.
+- It's to be used by automated unit tests in tests.py that can't connect to CWB in certain circumstances (e.g. when run on GitHub Action CI)
